+++ internal/home/rdns.go
@@ -106,8 +106,7 @@ func (r *RDNS) Begin(ip net.IP) {
 		return
 	}
 
-	id := ip.String()
-	if r.clients.Exists(id, ClientSourceRDNS) {
+	if r.clients.Exists(ip, ClientSourceRDNS) {
 		return
 	}
 
@@ -138,6 +137,6 @@ func (r *RDNS) workerLoop() {
 
 		// Don't handle any errors since AddHost doesn't return non-nil
 		// errors for now.
-		_, _ = r.clients.AddHost(ip.String(), host, ClientSourceRDNS)
+		_, _ = r.clients.AddHost(ip, host, ClientSourceRDNS)
 	}
 }